+++ src/components/ListEntry.vue
@@ -6,7 +6,7 @@
 
 <script>
 export default {
-  name: 'HelloWorld',
+  name: 'ListEntry',
   props: {
     class: 
   }
